+++ ApiTP1MethodeDeDev/LoanRequestTest/Tests.csproj
@@ -21,4 +21,8 @@
     <Using Include="Microsoft.VisualStudio.TestTools.UnitTesting" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="BorrowerTest\" />
+  </ItemGroup>
+
 </Project>
